refactor: rm old drawer and hardcoded load params
