--- frontend/core/containers/tool/Drawer/Header/CloseLine.tsx
@@ -1,21 +1,0 @@
-import type { FC } from 'react'
-
-import useLogic from '../useLogic'
-import useSalon from '../salon/header/close_line'
-
-type TProps = {
-  curve: boolean
-}
-
-const CloseLine: FC<TProps> = ({ curve }) => {
-  const { closeDrawer } = useLogic()
-  const s = useSalon()
-
-  return (
-    <div className={s.wrapper} onClick={() => closeDrawer()}>
-      <div>todo</div>
-    </div>
-  )
-}
-
-export default CloseLine
