--- frontend/core/containers/tool/Drawer/salon/header/close_line.ts
@@ -1,11 +1,0 @@
-// ref see example: https://codepen.io/mattbraun/pen/EywBJR
-
-import useTwBelt from '~/hooks/useTwBelt'
-
-export default () => {
-  const { cn } = useTwBelt()
-
-  return {
-    wrapper: cn('relative mr-28 top-4 rotate-180'),
-  }
-}
