--- frontend/core/containers/tool/Drawer/tests/index.test.ts
@@ -1,10 +1,0 @@
-// import React from 'react'
-// import { shallow } from 'enzyme'
-
-// import Preview from '..'
-
-describe('<Preview />', () => {
-  it('Expect to have unit tests specified', () => {
-    expect(true).toEqual(true)
-  })
-})
