--- frontend/core/containers/tool/Drawer/tests/store.test.ts
@@ -1,9 +1,0 @@
-/*
- * DrawerStore store test
- *
- */
-// import DrawerStore from '..'
-
-it('1 + 1 = 2', () => {
-  expect(1 + 1).toBe(2)
-})
