+++ frontend/core/containers/viewer/ArticleViewer/index.tsx
@@ -1,3 +1,4 @@
+'use client'
 /*
  *
  * ArticleViewer
@@ -5,22 +6,25 @@
  */
 import { useEffect } from 'react'
 import EVENT from '~/const/event'
-// import Comments from '~/containers/unit/Comments'
+import Comments from '~/containers/unit/Comments'
 import { send } from '~/signal'
 import LavaLampLoading from '~/widgets/Loading/LavaLampLoading'
 
+import type { TArticleLoad } from '~/spec'
 import DrawerHeader from './DrawerHeader'
 import useSalon from './salon'
 
 import useLogic from './useLogic'
 import Viewer from './Viewer'
 
-export default () => {
+type TProps = TArticleLoad 
+
+export default (props: TProps) => {
   const s = useSalon()
   const { article, loadArticle } = useLogic()
 
   useEffect(() => {
-    loadArticle()
+    loadArticle(props)
   }, [])
 
   useEffect(() => {
@@ -37,8 +41,7 @@ export default () => {
       <Viewer article={article} />
 
       <div className={s.comments}>
-        <h3>TODO: comments (slow down the drawer)</h3>
-        {/* <Comments /> */}
+        <Comments />
       </div>
     </div>
   )
