+++ frontend/core/providers/helper.ts
@@ -1,4 +1,4 @@
-import { cacheLife, unstable_cacheTag as cacheTag } from 'next/cache'
+import { cacheLife, cacheTag } from 'next/cache'
 import { isEmpty, mergeRight, reject } from 'ramda'
 import { CACHE_TAG } from '~/const/cache'
 import { LOCALE } from '~/const/i18n'
