+++ frontend/core/widgets/@Drawer/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { useRouter } from 'next/navigation'
 import { useEffect, useRef, useState } from 'react'
