+++ frontend/core/widgets/GlobalLayout/Addon.tsx
@@ -6,8 +6,6 @@ import { Toaster } from 'sonner'
 
 import useShortcut from '~/hooks/useShortcut'
 
-import Drawer from '~/containers/tool/Drawer'
-
 import PriceWall from '~/widgets/PriceWall/Modal'
 import UserListModal from '~/widgets/UserListModal'
 
@@ -18,7 +16,6 @@ const Addon: FC = () => {
 
   return (
     <>
-      <Drawer />
       <PriceWall />
       <UserListModal />
       <Toaster position="top-center" />
