+++ frontend/main/src/app/[community]/(articles)/post/@previewer/(.)[id]/page.tsx
@@ -1,20 +1,17 @@
-'use client'
-
-import dynamic from 'next/dynamic'
 import { Suspense } from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
 import ArticleViewer from '~/containers/viewer/ArticleViewer'
-// import Drawer from '~/widgets/@Drawer'
+import Drawer from '~/widgets/@Drawer'
 import LavaLampLoading from '~/widgets/Loading/LavaLampLoading'
 
-const Drawer = dynamic(() => import('~/widgets/@Drawer'), { ssr: false })
+export default async function Page({ params }) {
+  const p = await params
 
-export default function Page() {
   return (
     <Drawer>
       <ErrorBoundary fallback={<div>Error loading article</div>}>
         <Suspense fallback={<LavaLampLoading />}>
-          <ArticleViewer />
+          <ArticleViewer community={p.community} innerId={p.id} thread='post' />
         </Suspense>
       </ErrorBoundary>
     </Drawer>
